bot/requirements.txt
SQLAlchemy==2.0.12
aiogram==3.0.0b7
python-dotenv==1.0.0
asyncpg==0.27.0
cryptography==40.0.2